+++ src/CosmoFuse/correlations.py
@@ -17,7 +17,6 @@ from .utils import pixel2RaDec
 logger = logging.getLogger(__name__)
 
 _ALLOWED_FLOAT_PRECISIONS = {
-    "float16": np.float16,
     "float32": np.float32,
     "float64": np.float64,
 }
@@ -26,7 +25,6 @@ _ALLOWED_INDEX_PRECISIONS = {
     "uint64": np.uint64,
 }
 _ROTATION_COMPLEX_PRECISION = {
-    "float16": np.complex64,
     "float32": np.complex64,
     "float64": np.complex128,
 }
@@ -95,7 +93,7 @@ class Correlation:
         device: Union[str, int] = "auto",
         multiprocessing_start_method: str = "spawn",
         map_precision: Union[str, np.dtype, type] = "float64",
-        rotation_precision: Union[str, np.dtype, type] = "float64",
+        rotation_precision: Union[str, np.dtype, type] = "float32",
         index_precision: Union[str, np.dtype, type] = "uint32",
     ) -> None:
         """Initialize the Correlation class with validation.
@@ -115,8 +113,8 @@ class Correlation:
             multiprocessing_start_method: Start method used by multiprocessing when
                 pair calculations are parallelized. Use one of available Python start
                 methods or "default".
-            map_precision: One of float16/float32/float64 for map-like arrays.
-            rotation_precision: One of float16/float32/float64 for rotation/filter arrays.
+            map_precision: One of float32/float64 for map-like arrays.
+            rotation_precision: One of float32/float64 for rotation/filter arrays.
             index_precision: One of uint32/uint64 for index-like arrays.
 
         Raises:
@@ -340,7 +338,7 @@ class Correlation:
 
         return cos_2phi, sin_2phi, Q
 
-    def calculate_pairs_M_a(self, threads=1):
+    def calculate_pairs_M_a(self):
         self.Q_cos, self.Q_sin, self.Q_val, self.Q_inds, self.Q_patch_area = (
             [],
             [],
@@ -349,7 +347,7 @@ class Correlation:
             [],
         )
 
-        for i in range(self.n_patches):
+        for i in trange(self.n_patches):
             vec = hp.ang2vec(self.theta_center[i], self.phi_center[i])
             pix_center = hp.ang2pix(
                 self.nside, self.theta_center[i], self.phi_center[i]
@@ -378,7 +376,7 @@ class Correlation:
         Calculates the pairs and their angles for all patches for 2PCF & aperture mass.
         """
         logger.info("Calculating pairs for aperture mass")
-        self.calculate_pairs_M_a(threads)
+        self.calculate_pairs_M_a()
         logger.info("Calculating pairs for 2PCF")
         self.calculate_pairs_2PCF(threads)
 
@@ -508,36 +506,6 @@ class Correlation:
         self.ntotpairs = size
         self._prepare_version += 1
 
-    def load_maps(
-        self, g11, g21, g12, g22, w1, w2, sumofweights=None, flip_g1=True, flip_g2=False
-    ):
-        if self.inds_dev is None:
-            self.prepare()
-
-        self.g11 = self.backend.to_device(np.asarray(g11, dtype=self.map_dtype))
-        self.g21 = self.backend.to_device(np.asarray(g21, dtype=self.map_dtype))
-        self.g12 = self.backend.to_device(np.asarray(g12, dtype=self.map_dtype))
-        self.g22 = self.backend.to_device(np.asarray(g22, dtype=self.map_dtype))
-        self.w1 = self.backend.to_device(np.asarray(w1, dtype=self.map_dtype))
-        self.w2 = self.backend.to_device(np.asarray(w2, dtype=self.map_dtype))
-
-        if sumofweights is None:
-            self.sumofweights = self.backend.add.reduceat(
-                self.w1[self.inds_dev[0]] * self.w2[self.inds_dev[1]],
-                self.tot_bins_reduceat_dev[:-1],
-            )
-        else:
-            self.sumofweights = self.backend.to_device(
-                np.asarray(sumofweights, dtype=self.map_dtype)
-            )
-
-        if flip_g1:
-            self.g11 = -self.g11
-            self.g12 = -self.g12
-        if flip_g2:
-            self.g21 = -self.g21
-            self.g22 = -self.g22
-
     def xipm(self, g11, g21, g12, g22, w1, w2, sumofweights):
         # Arrays should already be on device if passed correctly, or we assume they are.
         # But xipm is called with slices often, which are views.
@@ -574,28 +542,6 @@ class Correlation:
 
         # Real parts
         return xip.real, xim.real
-
-    def get_all_xipm(self):
-        xip1, xim1 = self.xipm(
-            self.g11,
-            self.g21,
-            self.g12,
-            self.g22,
-            self.w1,
-            self.w2,
-            self.sumofweights,
-        )
-        xip2, xim2 = self.xipm(
-            self.g12,
-            self.g22,
-            self.g11,
-            self.g21,
-            self.w2,
-            self.w1,
-            self.sumofweights,
-        )
-
-        return (xip1 + xip2) / 2, (xim1 + xim2) / 2
 
     def _fingerprint_weights(self, w_np):
         w_contiguous = np.ascontiguousarray(w_np)
@@ -685,11 +631,11 @@ class Correlation:
             self._tomo_sumofweights_cache_w_fingerprint = w_fingerprint
             self._tomo_sumofweights_cache_prepare_version = self._prepare_version
 
-        g1, g2 = 1, 1
+        g1_fac, g2_fac = 1, 1
         if flip_g1:
-            g1 = -1
+            g1_fac = -1
         if flip_g2:
-            g2 = -1
+            g2_fac = -1
 
         M_ap = np.zeros([nzbins, self.n_patches], dtype=self.map_dtype)
         map_backend_dtype = getattr(self.backend.module, self.map_dtype.name)
@@ -710,26 +656,21 @@ class Correlation:
 
         k = 0
         for i in range(nzbins):
-            # M_a calculation remains on CPU as it uses numba njit on CPU arrays typically?
+            # M_a calculation remains on CPU as it uses numba njit on CPU arrays
             # get_M_a uses self.M_A_patch which is a njit function.
-            # It expects CPU arrays? Let's check get_M_a.
-            # It uses self.Q_inds[i], self.Q_cos[i] etc which are lists of numpy arrays (CPU).
-            # So shear_maps and w passed to it must be CPU arrays.
-            # shear_maps[i, 0] is a numpy array view if shear_maps is numpy array.
-
-            # Note: shear_maps passed to this function is likely numpy array.
+            # shear_maps and w passed to it must be CPU arrays.
             # shear_maps_dev is the device copy.
 
             M_ap[i] = self.get_M_a(
-                g1 * shear_maps_np[i, 0], g2 * shear_maps_np[i, 1], w_np[i]
+                g1_fac * shear_maps_np[i, 0], g2_fac * shear_maps_np[i, 1], w_np[i]
             )
             for j in range(i, nzbins):
                 if i == j:
                     xip1[k], xim1[k] = self.xipm(
-                        g1 * shear_maps_dev[i, 0],
-                        g2 * shear_maps_dev[i, 1],
-                        g1 * shear_maps_dev[j, 0],
-                        g2 * shear_maps_dev[j, 1],
+                        g1_fac * shear_maps_dev[i, 0],
+                        g2_fac * shear_maps_dev[i, 1],
+                        g1_fac * shear_maps_dev[j, 0],
+                        g2_fac * shear_maps_dev[j, 1],
                         w_dev[i],
                         w_dev[j],
                         sumofweights_dev[0, k],
@@ -737,46 +678,28 @@ class Correlation:
                     xip2[k], xim2[k] = xip1[k], xim1[k]
                 else:
                     xip1[k], xim1[k] = self.xipm(
-                        g1 * shear_maps_dev[i, 0],
-                        g2 * shear_maps_dev[i, 1],
-                        g1 * shear_maps_dev[j, 0],
-                        g2 * shear_maps_dev[j, 1],
+                        g1_fac * shear_maps_dev[i, 0],
+                        g2_fac * shear_maps_dev[i, 1],
+                        g1_fac * shear_maps_dev[j, 0],
+                        g2_fac * shear_maps_dev[j, 1],
                         w_dev[i],
                         w_dev[j],
                         sumofweights_dev[0, k],
                     )
                     xip2[k], xim2[k] = self.xipm(
-                        g1 * shear_maps_dev[j, 0],
-                        g2 * shear_maps_dev[j, 1],
-                        g1 * shear_maps_dev[i, 0],
-                        g2 * shear_maps_dev[i, 1],
+                        g1_fac * shear_maps_dev[j, 0],
+                        g2_fac * shear_maps_dev[j, 1],
+                        g1_fac * shear_maps_dev[i, 0],
+                        g2_fac * shear_maps_dev[i, 1],
                         w_dev[j],
                         w_dev[i],
                         sumofweights_dev[1, k],
                     )
                 k += 1
-
+        
+        # pairs are just upper triangle, so average over both combinations
         xip = (xip1 + xip2) / 2
         xim = (xim1 + xim2) / 2
 
         # Return as numpy arrays
         return M_ap, self.backend.to_numpy(xip), self.backend.to_numpy(xim)
-
-    def calculate_2PCF(self, threads=1):
-        """
-        Calculates the 2PCF for the loaded maps.
-
-        This method computes xip and xim using the loaded maps (loaded via load_maps).
-        It mimics the behavior of the old Correlation_CPU.calculate_2PCF but uses the vectorized xipm implementation.
-        """
-        # Ensure data is prepared
-        if self.inds_dev is None:
-             self.prepare()
-
-        # If maps are not loaded, this will fail.
-        # But wait, calculate_2PCF is usually called after load_maps.
-
-        xip, xim = self.get_all_xipm()
-        self.xip = self.backend.to_numpy(xip)
-        self.xim = self.backend.to_numpy(xim)
-        return self.xip, self.xim